requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
joblib
matplotlib
seaborn
langchain
langchain-google-genai
langchain-community
faiss-cpu
sentence-transformers
duckduckgo-search
python-dotenv
